Add createSlidersState
